basic creation of example plot
Using c elegans example data I added a plot of a region of reads using Gviz. it's not interactive yet, but committing for now since I was able to figure out how to get something plotted! The example bam is too large for upload, so I will likely not have an example in the final version...