remove shinythemes
Decided to remove the shinythemes because it just adds an extra dependency and the default theme actually looks appropriate for the plot being created. I do still like slate theme, but just seems like something unnecessary for an app that is only run locally. If I upload it to a server then that version might get a theme.